feat: add multiple delete functions for various entities
- Implemented DeletePayment, DeleteTeacherLesson, DeleteAttendance, DeleteRole, DeleteExpense, DeleteTask, DeleteExam, DeleteRemediation, DeleteSynchronization, DeletePlanning, DeletePlanningSlot, DeleteSchoolYear, DeleteNotification, DeleteStudentClass, DeleteTransactionHistory, DeleteRemuneration, and DeleteLeave functions in the delete controller.
- Each function retrieves an ID from the request context, attempts to delete the corresponding entity from the database, logs success or error messages, and redirects to the appropriate list page.
- Updated models to use uint64 for IDs instead of uint for better compatibility with larger datasets.
- Modified related model fields and functions to ensure consistency across the application.

+++ models/payements.go
@@ -8,7 +8,7 @@ import (
 
 type Payment struct {
 	ID        uint    `gorm:"primaryKey"`
-	StudentID uint    `gorm:"not null"`             // Clé étrangère vers Student
+	StudentID uint64    `gorm:"not null"`             // Clé étrangère vers Student
 	Student   Student `gorm:"foreignKey:StudentID"` // Relation GORM
 	Price     float64 `gorm:"not null"`             // Montant payé
 	Type      string  `gorm:"size:50;not null"`     // Ex: "Inscription", "Mensualité"
